[Feat] basic enemy making completed
