+++ Assets/enemy/scrt_enemyattack.cs
@@ -2,10 +2,10 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class scrt_enemyattack : MonoBehaviour
+public class scrt_enemyAttack : MonoBehaviour
 {
-    float attack;
-    int enemycode; //0: dustpan
+    public float attack;
+    public int enemyCode;
 
     // Start is called before the first frame update
     void Start()
@@ -17,5 +17,13 @@ public class scrt_enemyattack : MonoBehaviour
     void Update()
     {
 
+    }
+
+    private void OnTriggerEnter(Collider col) //물리충돌을 사용하지 않는 충돌 사용
+    {
+        if (col.tag == "player")
+        {
+            col.gameObject.GetComponent<scrt_player>().Damage(attack);
+        }
     }
 }
