+++ ProjectSettings/TagManager.asset
@@ -3,7 +3,12 @@
 --- !u!78 &1
 TagManager:
   serializedVersion: 2
-  tags: []
+  tags:
+  - player
+  - enemy
+  - wall
+  - playerWeapon
+  - enemyWeapon
   layers:
   - Default
   - TransparentFX
